requirements.txt
aiofiles==24.1.0
anyio==4.10.0
argcomplete==3.6.2
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
bcrypt==4.3.0
beautifulsoup4==4.13.5
bleach==6.2.0
boto3==1.40.35
botocore==1.40.35
Brotli==1.1.0
cachetools==6.2.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
comm==0.2.3
contourpy==1.3.3
cryptography==46.0.1
cycler==0.12.1
dacite==1.9.2
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
dill==0.4.0
dnspython==2.8.0
executing==2.2.1
fastjsonschema==2.21.2
fiftyone==1.8.0
fiftyone-brain==0.21.3
fiftyone_db==1.3.0
fonttools==4.60.0
fqdn==1.5.1
ftfy==6.3.1
glob2==0.7
graphql-core==3.2.6
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
humanize==4.13.0
Hypercorn==0.17.3
hyperframe==6.1.0
idna==3.10
imageio==2.37.0
inflate64==1.0.3
ipykernel==6.30.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.2
json5==0.12.1
jsonlines==4.0.0
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.7
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kiwisolver==1.4.9
lark==1.2.2
lazy_loader==0.4
lia-web==0.2.3
MarkupSafe==3.0.2
matplotlib==3.10.6
matplotlib-inline==0.1.7
mistune==3.1.4
mongoengine==0.29.1
motor==3.6.1
multivolumefile==0.2.3
narwhals==2.5.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
notebook==7.4.5
notebook_shim==0.2.4
numpy==2.2.6
opencv-python-headless==4.12.0.88
packaging==25.0
pandas==2.3.2
pandocfilters==1.5.1
paramiko==3.5.1
parso==0.8.5
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.4.0
plotly==6.3.0
pprintpp==0.4.0
priority==2.0.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
psutil==7.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
py7zr==1.0.0
pybcj==1.0.6
pycparser==2.23
pycryptodomex==3.23.0
pydash==8.0.5
Pygments==2.19.2
pymongo==4.9.2
PyNaCl==1.6.0
pyparsing==3.2.4
pyppmd==1.2.0
python-dateutil==2.9.0.post0
python-json-logger==3.3.0
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.1.0
pyzstd==0.17.0
rarfile==4.2
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
retrying==1.4.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.27.1
rtree==1.4.1
s3transfer==0.14.0
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.16.2
Send2Trash==1.8.3
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
sse-starlette==0.10.3
sseclient-py==1.8.0
stack-data==0.6.3
starlette==0.48.0
strawberry-graphql==0.282.0
tabulate==0.9.0
terminado==0.18.1
texttable==1.7.0
threadpoolctl==3.6.0
tifffile==2025.9.20
tinycss2==1.4.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20250822
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
universal-analytics-python3==1.1.1
uri-template==1.3.0
urllib3==2.5.0
voxel51-eta==0.15.1
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.14
wrapt==1.17.3
wsproto==1.2.0
xmltodict==1.0.2
